final part 2

+++ model/reviws-model.php
@@ -62,16 +62,62 @@ function getReviewsByClientId($clientId){
     return $reviews;
 }
 
-function getReview(){
+function getReview($reviewId){
+    // Create a connection object using the phpmotors connection function
+    $db = phpmotorsConnect();
+    // The SQL statement
+    $sql = 'SELECT * FROM reviews WHERE reviewId = :reviewId';
 
+    $stmt = $db->prepare($sql);
+
+    $stmt->bindValue(':reviewId', $reviewId, PDO::PARAM_INT);
+    // Insert the data
+    $stmt->execute();
+    // Ask how many rows changed as a result of our insert
+    $reviews = $stmt->fetch(PDO::FETCH_ASSOC);;
+    // Close the database interaction
+    $stmt->closeCursor();
+    // Return the indication of success (rows changed)
+    return $reviews;
 }
 
-function updateReview(){
+function updateReview($reviewId, $reviewText){
+    // Create a connection object using the phpmotors connection function
+    $db = phpmotorsConnect();
+    // The SQL statement
+    $sql = 'UPDATE reviews SET reviewText = :reviewText WHERE reviewId = :reviewId';
 
+    $stmt = $db->prepare($sql);
+
+    $stmt->bindValue(':reviewId', $reviewId, PDO::PARAM_INT);
+    $stmt->bindValue(':reviewText', $reviewText, PDO::PARAM_STR);
+    // Insert the data
+    $stmt->execute();
+    // Ask how many rows changed as a result of our insert
+    $rowsChanged = $stmt->rowCount();
+    // Close the database interaction
+    $stmt->closeCursor();
+    // Return the indication of success (rows changed)
+    return $rowsChanged;
 }
 
-function deleteReview(){
+function deleteReview($reviewId){
+    // Create a connection object using the phpmotors connection function
+    $db = phpmotorsConnect();
+    // The SQL statement
+    $sql = 'DELETE FROM reviews WHERE reviewId = :reviewId';
 
+    $stmt = $db->prepare($sql);
+
+    $stmt->bindValue(':reviewId', $reviewId, PDO::PARAM_INT);
+    // Insert the data
+    $stmt->execute();
+    // Ask how many rows changed as a result of our insert
+    $rowsChanged = $stmt->rowCount();
+    // Close the database interaction
+    $stmt->closeCursor();
+    // Return the indication of success (rows changed)
+    return $rowsChanged;
 }
 
 function buildReviewView($reviewList){
@@ -95,7 +141,7 @@ function buildReviewAdmin($reviewList)
         $reviews .= "<h3>" . $review['clientFirstname'][0] . $review['clientLastname'] . "</h3>";
         $reviews .= "<p>" . $review['reviewText'] . "</p>";
         $reviews .= $review['reviewDate'] . "<br>";
-        $reviews .= "<a href='/phpmotors/reviews/index.php?action=edit'>Edit</a> <a href='/phpmotors/reviews/index.php?action=delete'>Delete</a>";
+        $reviews .= "<a href='/phpmotors/reviews/index.php?action=edit&reviewId=" . $review['reviewId'] . "'>Edit</a> <a href='/phpmotors/reviews/index.php?action=confirm-delete&reviewId=". $review['reviewId']."'>Delete</a>";
         $reviews .= "</div>";
     }
     $reviews .= "</div>";
